Add conformance tests for device frame and wire ABI
- Introduced `device_frame.rs` to validate frame preflight assertions against the frozen `conformance/v1.0` golden vectors.
- Added tests to ensure valid frames are dispatchable, malformed frames write exact error headers, and unusable frames do not become dispatchable.
- Created `wire_abi_conformance.rs` to assert wire-ABI compliance with the frozen `conformance/v1.0` golden artifacts.
- Implemented layout assertions for various wire structures to match the versioned manifest.
- Added tests for canonical frames, edge cases, and independent codec interoperability with primary wire types.
