+++ .github/workflows/ci.yml
@@ -184,6 +184,28 @@ jobs:
       - name: Check portable dependency features
         run: bash ci/check-portable-dependencies.sh
 
+  publish-dry-run:
+    name: publish-dry-run
+    runs-on: ubuntu-latest
+    timeout-minutes: 30
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+        with:
+          persist-credentials: false
+
+      - name: Install stable Rust
+        uses: dtolnay/rust-toolchain@fa04a1451ff1842e2626ccb99004d0195b455a88
+        with:
+          toolchain: stable
+
+      # Packages every published crate in documented order into an isolated local registry, then
+      # builds, tests, and documents each one from its extracted tarball. `cargo package`'s own
+      # verify step only builds the library target, which is how four cross-package `include_str!`
+      # sites went unnoticed; this job runs the tests inside the packaged source instead.
+      - name: Run the ordered publication dry run
+        run: python3 ci/publish-dry-run.py
+
   dependency-policy:
     name: dependency-policy
     runs-on: ubuntu-latest
