+++ Cargo.toml
@@ -4,8 +4,24 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
 description = "Portable Rust foundations for a virtio accelerator device"
-publish = false
+readme = "README.md"
+include = [
+    "/Cargo.toml",
+    "/README.md",
+    "/SECURITY.md",
+    "/LICENSE-MIT",
+    "/LICENSE-APACHE",
+    "/src/**",
+    "/examples/**",
+    "/tests/**",
+    "/docs/**",
+    "/conformance/v1.0/**",
+]
 
 [workspace]
 members = [
@@ -26,20 +42,24 @@ version = "0.1.0"
 edition = "2024"
 license = "MIT OR Apache-2.0"
 rust-version = "1.85"
+repository = "https://github.com/MicroPerceptron/virtio-accel"
+homepage = "https://github.com/MicroPerceptron/virtio-accel"
+keywords = ["virtio", "accelerator", "npu", "no-std", "protocol"]
+categories = ["no-std", "hardware-support", "virtualization"]
 
 [workspace.dependencies]
 bitflags = { version = "2.13.1", default-features = false }
 serde_json = "1.0.149"
 zerocopy = { version = "0.8.54", default-features = false, features = ["derive"] }
-virtio-accel-core = { path = "crates/virtio-accel-core" }
-virtio-accel-cleanroom = { path = "conformance/rust-clean-room" }
-virtio-accel-conformance = { path = "crates/virtio-accel-conformance" }
-virtio-accel-device = { path = "crates/virtio-accel-device" }
-virtio-accel-guest = { path = "crates/virtio-accel-guest" }
-virtio-accel-mock = { path = "crates/virtio-accel-mock" }
-virtio-accel-proto = { path = "crates/virtio-accel-proto" }
-virtio-accel-split-queue = { path = "crates/virtio-accel-split-queue" }
-virtio-accel-transport = { path = "crates/virtio-accel-transport" }
+virtio-accel-core = { path = "crates/virtio-accel-core", version = "0.1.0" }
+virtio-accel-cleanroom = { path = "conformance/rust-clean-room", version = "0.1.0" }
+virtio-accel-conformance = { path = "crates/virtio-accel-conformance", version = "0.1.0" }
+virtio-accel-device = { path = "crates/virtio-accel-device", version = "0.1.0" }
+virtio-accel-guest = { path = "crates/virtio-accel-guest", version = "0.1.0" }
+virtio-accel-mock = { path = "crates/virtio-accel-mock", version = "0.1.0" }
+virtio-accel-proto = { path = "crates/virtio-accel-proto", version = "0.1.0" }
+virtio-accel-split-queue = { path = "crates/virtio-accel-split-queue", version = "0.1.0" }
+virtio-accel-transport = { path = "crates/virtio-accel-transport", version = "0.1.0" }
 
 [dependencies]
 virtio-accel-core.workspace = true
@@ -51,5 +71,7 @@ virtio-accel-transport.workspace = true
 
 [dev-dependencies]
 serde_json.workspace = true
+virtio-accel-cleanroom.workspace = true
 virtio-accel-conformance.workspace = true
 virtio-accel-mock.workspace = true
+zerocopy.workspace = true
