+++ conformance/rust-clean-room/Cargo.toml
@@ -4,8 +4,10 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
 description = "Independent manual codec for virtio-accel protocol conformance"
-publish = false
+readme = "README.md"
 
-[dev-dependencies]
-serde_json.workspace = true
