+++ crates/virtio-accel-core/Cargo.toml
@@ -4,8 +4,12 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
 description = "Transport-independent accelerator lifecycle traits"
-publish = false
+readme = "README.md"
 
 [dependencies]
 bitflags.workspace = true
