+++ crates/virtio-accel-device/Cargo.toml
@@ -4,8 +4,12 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
 description = "Transport-neutral virtio accelerator device state"
-publish = false
+readme = "README.md"
 
 [dependencies]
 virtio-accel-core.workspace = true
@@ -14,5 +18,4 @@ virtio-accel-transport.workspace = true
 zerocopy.workspace = true
 
 [dev-dependencies]
-serde_json.workspace = true
 virtio-accel-mock.workspace = true
