+++ crates/virtio-accel-mock/Cargo.toml
@@ -4,8 +4,12 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
 description = "Cross-platform reference backend for virtio-accel"
-publish = false
+readme = "README.md"
 
 [dependencies]
 virtio-accel-core.workspace = true
