+++ crates/virtio-accel-proto/Cargo.toml
@@ -4,13 +4,16 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
 description = "Pointer-free portable virtio accelerator protocol 1.0 ABI"
-publish = false
+readme = "README.md"
 
 [dependencies]
 bitflags.workspace = true
 zerocopy.workspace = true
 
 [dev-dependencies]
-serde_json.workspace = true
 virtio-accel-cleanroom.workspace = true
