+++ crates/virtio-accel-split-queue/Cargo.toml
@@ -4,8 +4,12 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
 description = "Bounded in-memory split-virtqueue reference model"
-publish = false
+readme = "README.md"
 
 [dependencies]
 virtio-accel-transport.workspace = true
