+++ crates/virtio-accel-transport/Cargo.toml
@@ -4,7 +4,11 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
 description = "Portable descriptor-chain and virtqueue port contracts"
-publish = false
+readme = "README.md"
 
 [dependencies]
